feat(data-masking): move AWSEncryptionSDKProvider to providers/kms sub-path
- provider now lives at @aws-lambda-powertools/data-masking/providers/kms,
  following the Idempotency persistence-layer pattern: the optional
  @aws-crypto/client-node peer dependency is opt-in via import path
- dynamic import() replaced with static top-level imports; #init removed
  and client assembly moved into the constructor
- provider options and SDK client types extracted to providers/kms/types.ts
  with a providers/kms/types sub-path export
- docs and e2e function code updated to the new import path

Addresses review comments on #5143
