--- packages/data-masking/src/provider/index.ts
@@ -1,4 +1,0 @@
-export {
-  AWSEncryptionSDKProvider,
-  type AWSEncryptionSDKProviderOptions,
-} from './AWSEncryptionSDKProvider.js';
