+++ packages/data-masking/src/providers/kms/index.ts
@@ -1,0 +1,2 @@
+export { AWSEncryptionSDKProvider } from './AWSEncryptionSDKProvider.js';
+export type { AWSEncryptionSDKProviderOptions } from './types.js';
